requirements.txt
fastapi[standard]==0.121.2
sqlalchemy==2.0.44
pydantic_settings==2.12.0
psycopg2==2.9.11
asyncpg==0.30.0
alembic==1.17.2
passlib[argon2]==1.7.4
pyjwt==2.10.1
black==25.9.0
isort==7.0.0